+++ history.md
@@ -100,7 +100,8 @@
   - *breaking:* `ServiceContainer.GetEnumerator()` no longer clones the container (enumerating is still thread safe)
   - *breaking:* dropped `netstandard1.6` support
   - *breaking:* new `IComposite` layout
-  - *breaking:* `AbstractServiceEntry` moved to public API
+  - *breaking:* `AbstractServiceEntry` moved to the public API
+  - *breaking:* `ServiceReference` moved to the public API
   - *added:* `IAsyncDisposable` support
   - *done:* using nullable reference types
   - *fixed:* named services could not be queried via `ServiceProvider` ([details](https://github.com/Sholtee/injector/commit/bd97a0caf2763b3a28d4c511b2ee02006b072b3b ))